+++ src/main/java/io/dutwrapperlib/dutwrapper/model/enums/NewsType.java
@@ -1,0 +1,6 @@
+package io.dutwrapperlib.dutwrapper.model.enums;
+
+public enum NewsType {
+    Global,
+    Subject
+}