+++ src/test/java/io/dutwrapperlib/dutwrapper/NewsGlobalTest.java
@@ -2,8 +2,8 @@ package io.dutwrapperlib.dutwrapper;
 
 import org.junit.jupiter.api.Test;
 
-import io.dutwrapperlib.dutwrapper.objects.news.LinkItem;
-import io.dutwrapperlib.dutwrapper.objects.news.NewsGlobalItem;
+import io.dutwrapperlib.dutwrapper.model.news.LinkItem;
+import io.dutwrapperlib.dutwrapper.model.news.NewsGlobalItem;
 
 import java.util.List;
 
@@ -19,14 +19,16 @@ class NewsGlobalTest {
             System.out.println(String.format("Page %d (%d item(s))", page, newsList.size()));
 
             for (NewsGlobalItem newsItem : newsList) {
-                System.out.println(String.format("Index %d (date: %d)", newsList.indexOf(newsItem) + 1, newsItem.getDate()));
+                System.out.println(
+                        String.format("Index %d (date: %d)", newsList.indexOf(newsItem) + 1, newsItem.getDate()));
                 System.out.println(newsItem.getTitle());
                 System.out.println(newsItem.getContentString());
                 System.out.println(String.format("%d link(s)", newsItem.getLinks().size()));
-                if (newsItem.getLinks().size() > 0 ) {
+                if (newsItem.getLinks().size() > 0) {
                     System.out.println("Position | Text | URL");
                     for (LinkItem linkItem : newsItem.getLinks()) {
-                        System.out.println(String.format("%d | %s | %s", linkItem.getPosition(), linkItem.getText(), linkItem.getUrl()));
+                        System.out.println(String.format("%d | %s | %s", linkItem.getPosition(), linkItem.getText(),
+                                linkItem.getUrl()));
                     }
                 }
             }